# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MentorAlm API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupInput {
  name: string;
  email: string;
  phone?: string;
  /** @minLength 6 */
  password: string;
  confirmPassword: string;
}

export interface SigninInput {
  email: string;
  password: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  email: string;
  otp: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface UserPublic {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  accessToken: string;
  user: UserPublic;
}

export interface RefreshResponse {
  accessToken: string;
}

export interface StudentProfile {
  id: number;
  userId: number;
  dateOfBirth?: string | null;
  gender?: string | null;
  city?: string | null;
  state?: string | null;
  photoUrl?: string | null;
  educationLevel?: string | null;
  board?: string | null;
  schoolCollege?: string | null;
  gradePercentage?: string | null;
  stream?: string | null;
  interests?: string | null;
  strengths?: string | null;
  hobbies?: string | null;
  energizedBy?: string | null;
  dreamCareer?: string | null;
  targetColleges?: string | null;
  openToAbroad?: string | null;
  biggestConcern?: string | null;
  familyIncome?: string | null;
  parentsEducation?: string | null;
  familyPressure?: string | null;
  educationBudget?: string | null;
  fiveYearGoal?: string | null;
  alreadyTried?: string | null;
  stoppingYou?: string | null;
  heardFrom?: string | null;
  completionPercent: number;
  updatedAt: string;
  createdAt: string;
}

export interface UpdateProfileInput {
  dateOfBirth?: string;
  gender?: string;
  city?: string;
  state?: string;
  educationLevel?: string;
  board?: string;
  schoolCollege?: string;
  gradePercentage?: string;
  stream?: string;
  interests?: string;
  strengths?: string;
  hobbies?: string;
  energizedBy?: string;
  dreamCareer?: string;
  targetColleges?: string;
  openToAbroad?: string;
  biggestConcern?: string;
  familyIncome?: string;
  parentsEducation?: string;
  familyPressure?: string;
  educationBudget?: string;
  fiveYearGoal?: string;
  alreadyTried?: string;
  stoppingYou?: string;
  heardFrom?: string;
}

export interface PhotoUploadResponse {
  photoUrl: string;
}

export interface Service {
  id: number;
  title: string;
  shortDesc: string;
  fullDesc: string;
  included: string;
  category: string;
  duration: string;
  price: number;
  counsellorName: string;
  counsellorBio: string;
  thumbnailUrl?: string | null;
  status: string;
  slots: string;
  createdAt: string;
}

export interface CreateServiceInput {
  title: string;
  shortDesc: string;
  fullDesc: string;
  included: string;
  category: string;
  duration: string;
  price: number;
  counsellorName: string;
  counsellorBio: string;
  thumbnailUrl?: string;
  status?: string;
  slots?: string;
}

export interface UpdateServiceInput {
  title?: string;
  shortDesc?: string;
  fullDesc?: string;
  included?: string;
  category?: string;
  duration?: string;
  price?: number;
  counsellorName?: string;
  counsellorBio?: string;
  thumbnailUrl?: string;
  status?: string;
  slots?: string;
}

export interface CreateBookingInput {
  serviceId: number;
  slotDateTime: string;
}

export interface VerifyPaymentInput {
  bookingId: number;
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
}

export interface Booking {
  id: number;
  userId: number;
  serviceId: number;
  slotDateTime: string;
  paymentStatus: string;
  razorpayOrderId?: string | null;
  razorpayPaymentId?: string | null;
  amount: number;
  status: string;
  createdAt: string;
}

export interface BookingWithService {
  id: number;
  userId: number;
  serviceId: number;
  slotDateTime: string;
  paymentStatus: string;
  razorpayOrderId?: string | null;
  razorpayPaymentId?: string | null;
  amount: number;
  status: string;
  createdAt: string;
  service: Service;
}

export interface BookingOrderResponse {
  bookingId: number;
  razorpayOrderId: string;
  razorpayKeyId: string;
  amount: number;
  currency: string;
}

export interface ChatMessage {
  id: number;
  userId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatMessageInput {
  content: string;
}

export interface Roadmap {
  id: number;
  userId: number;
  content: string;
  generatedAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  message: string;
  createdAt: string;
}

export interface AdminStats {
  totalStudents: number;
  activeThisWeek: number;
  totalRevenue: number;
  bookingsThisMonth: number;
}

export interface AdminStudent {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  completionPercent: number;
  createdAt: string;
}

export interface AdminStudentDetail {
  user: UserPublic;
  profile?: StudentProfile;
}

export interface AdminBooking {
  id: number;
  userId: number;
  studentName: string;
  studentEmail: string;
  serviceId: number;
  serviceTitle: string;
  slotDateTime: string;
  paymentStatus: string;
  amount: number;
  status: string;
  createdAt: string;
}

export interface UpdateBookingStatusInput {
  status: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type ListServicesParams = {
category?: string;
};

export type GetChatMessagesParams = {
limit?: number;
};

export type ListAdminBookingsParams = {
serviceId?: number;
status?: string;
};