# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MentorAlm API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List published services
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking and Razorpay order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Booking initiated with Razorpay order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingOrderResponse"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/students:
    get:
      operationId: listAdminStudents
      tags: [admin]
      summary: List all students
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminStudent"
  /admin/students/{id}:
    get:
      operationId: getAdminStudent
      tags: [admin]
      summary: Get full student profile (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full student data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudentDetail"
  /admin/bookings:
    get:
      operationId: listAdminBookings
      tags: [admin]
      summary: List all bookings
      security:
        - bearerAuth: []
      parameters:
        - name: serviceId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminBooking"
  /admin/bookings/{id}:
    patch:
      operationId: updateAdminBooking
      tags: [admin]
      summary: Update booking status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusInput"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBooking"
  /admin/services:
    post:
      operationId: createAdminService
      tags: [admin]
      summary: Create a service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /admin/services/{id}:
    patch:
      operationId: updateAdminService
      tags: [admin]
      summary: Update a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceInput"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteAdminService
      tags: [admin]
      summary: Delete a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/contacts:
    get:
      operationId: listAdminContacts
      tags: [admin]
      summary: List contact form submissions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dateOfBirth:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        educationLevel:
          type: string
          nullable: true
        board:
          type: string
          nullable: true
        schoolCollege:
          type: string
          nullable: true
        gradePercentage:
          type: string
          nullable: true
        stream:
          type: string
          nullable: true
        interests:
          type: string
          nullable: true
        strengths:
          type: string
          nullable: true
        hobbies:
          type: string
          nullable: true
        energizedBy:
          type: string
          nullable: true
        dreamCareer:
          type: string
          nullable: true
        targetColleges:
          type: string
          nullable: true
        openToAbroad:
          type: string
          nullable: true
        biggestConcern:
          type: string
          nullable: true
        familyIncome:
          type: string
          nullable: true
        parentsEducation:
          type: string
          nullable: true
        familyPressure:
          type: string
          nullable: true
        educationBudget:
          type: string
          nullable: true
        fiveYearGoal:
          type: string
          nullable: true
        alreadyTried:
          type: string
          nullable: true
        stoppingYou:
          type: string
          nullable: true
        heardFrom:
          type: string
          nullable: true
        completionPercent:
          type: number
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - completionPercent
        - updatedAt
        - createdAt
...
    Service:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        shortDesc:
          type: string
        fullDesc:
          type: string
        included:
          type: string
        category:
          type: string
        duration:
          type: string
        price:
          type: integer
        counsellorName:
          type: string
        counsellorBio:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        status:
          type: string
        slots:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - shortDesc
        - fullDesc
        - included
        - category
        - duration
        - price
        - counsellorName
        - counsellorBio
        - status
        - slots
        - createdAt
    CreateServiceInput:
      type: object
      properties:
        title:
          type: string
        shortDesc:
          type: string
        fullDesc:
          type: string
        included:
          type: string
        category:
          type: string
        duration:
          type: string
        price:
          type: integer
        counsellorName:
          type: string
        counsellorBio:
          type: string
        thumbnailUrl:
          type: string
        status:
          type: string
        slots:
          type: string
      required:
        - title
        - shortDesc
        - fullDesc
        - included
        - category
        - duration
        - price
        - counsellorName
        - counsellorBio
    UpdateServiceInput:
      type: object
      properties:
        title:
          type: string
        shortDesc:
          type: string
        fullDesc:
          type: string
        included:
          type: string
        category:
          type: string
        duration:
          type: string
        price:
          type: integer
        counsellorName:
          type: string
        counsellorBio:
          type: string
        thumbnailUrl:
          type: string
        status:
          type: string
        slots:
          type: string
    CreateBookingInput:
      type: object
      properties:
        serviceId:
          type: integer
        slotDateTime:
          type: string
      required:
        - serviceId
        - slotDateTime
...
    BookingOrderResponse:
      type: object
      properties:
        bookingId:
          type: integer
        razorpayOrderId:
          type: string
        razorpayKeyId:
          type: string
        amount:
          type: integer
        currency:
          type: string
      required:
        - bookingId
        - razorpayOrderId
        - razorpayKeyId
        - amount
        - currency
...
    ContactMessage:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - message
        - createdAt
    AdminStats:
      type: object
      properties:
        totalStudents:
          type: integer
        activeThisWeek:
          type: integer
        totalRevenue:
          type: integer
        bookingsThisMonth:
          type: integer
      required:
        - totalStudents
        - activeThisWeek
        - totalRevenue
        - bookingsThisMonth
    AdminStudent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        completionPercent:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - completionPercent
        - createdAt
    AdminStudentDetail:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserPublic"
        profile:
          $ref: "#/components/schemas/StudentProfile"
      required:
        - user
    AdminBooking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        studentName:
          type: string
        studentEmail:
          type: string
        serviceId:
          type: integer
        serviceTitle:
          type: string
        slotDateTime:
          type: string
        paymentStatus:
          type: string
        amount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - studentName
        - studentEmail
        - serviceId
        - serviceTitle
        - slotDateTime
        - paymentStatus
        - amount
        - status
        - createdAt
    UpdateBookingStatusInput:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...