# lib/api-zod/src/generated/types/adminStudent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MentorAlm API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStudent {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  completionPercent: number;
  createdAt: Date;
}