# lib/api-zod/src/generated/types/updateBookingStatusInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MentorAlm API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBookingStatusInput {
  status: string;
}