# lib/api-zod/src/generated/types/updateServiceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MentorAlm API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateServiceInput {
  title?: string;
  shortDesc?: string;
  fullDesc?: string;
  included?: string;
  category?: string;
  duration?: string;
  price?: number;
  counsellorName?: string;
  counsellorBio?: string;
  thumbnailUrl?: string;
  status?: string;
  slots?: string;
}